:bug: Fix deleting resources from Helm charts (#31)
There are two problems with deleting resources from Helm charts right
now: The first one is that the objects stay in the status.resources
list, even after they are deleted. The second is that instead of
returning an error when a deletion failed, we returned an error on
success.

Instead, as we base deletion of objects based on the status.resources
list (this can be changed in the future), we keep resources that are
supposed to be deleted but failed to in the status as not-synced.
Additionally, we requeue and try the deletion another time, just as we
do with applying currently.

As the current Update method is unused since we changed from create to
apply, it is deleted. Mocks are generated freshly as well.

Signed-off-by: janiskemper <janis.kemper@syself.com>