:bug: Fix delete of csr if assets cannot be downloaded
Fix delete of ClusterStackRelease object in case that there is a problem
with the release assets.

Currently, the reconcileDelete function requires assets to be downloaded
properly. If this is not the case, then the object does not get deleted
and is stuck forever.

This happens for example when a user specifies a version of a
clusterstack that does not exist. Then a ClusterStackRelease object is
created, but will never get deleted again, because the downloading of
the release assets fails, as the release does not exist.

Now we have an additional check to remove the finalizer without any
further logic in case there is an issue with the release assets.

Signed-off-by: janiskemper <janis.kemper@syself.com>