fix upgrade builder image script
when I initially worked on this, this was hardcoded
and we were using grep + sed combination to get
version. Later we switched to track version by
using a file at the root of the repo.
The file was .builder-image-version.txt, this was
a bug where we were still reading from Makefile
hardcoded version.
This commit fixes the same and switches to new
method where we read from .builder-image-version.txt

Signed-off-by: kranurag7 <anurag.kumar@syself.com>