use scs capd image while using docker provider
users on fedora using btrfs filesystem were having issues with capd and
that was fixed on main and it requires building your own image and using
the same here for now.

I tested it by building the image locally and now I don't see any error.
This will be fixed when we upgrade to 1.6.3 or later version.
As of now, we will rely on image built from CAPI main branch and pushed
to ghcr.io/sovereigncloudstack/capd with 1.0.5 tag.

Signed-off-by: kranurag7 <anurag.kumar@syself.com>