:seedling: Fix linting, tests and prepare for merge
Fix linting in new clusteraddon-related files that we changed in the
feature branch.

Fixing unit tests.

Prepare feature branch to merge by going through changes of PR and
validating them. Fixing some small stuff, as well as changing
unnecessary or temporary changes back to the state of the main branch.

Signed-off-by: janiskemper <janis.kemper@syself.com>